app/requirements.txt
PyMuPDF==1.23.7
layoutparser[ocr]
Pillow